# src/TripBundle/Resources/api/openapi.yaml
openapi: "3.0.5"
info:
    version: 1.0.0
    title: Trips
    license:
        name: GNU AGPLv3
servers:
    -   url: /
security:
    -   BasicAuth: [ ]
paths:
    /trip/accounts/:
        post:
            summary: Create account
            operationId: createAccount
            tags:
                - accounts
            requestBody:
                description: Data of new account
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AccountCreate'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Account'

            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Account"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
    /trip/accounts/{accountId}:
        get:
            summary: Get account information
            operationId: getAccount
            tags:
                - accounts
            parameters:
                -   name: accountId
                    in: path
                    required: true
                    schema:
                        type: integer
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Account"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
        patch:
            summary: Update account information
            operationId: updateAccount
            tags:
                - accounts
            parameters:
                -   name: accountId
                    in: path
                    required: true
                    schema:
                        type: integer
            requestBody:
                description: Updatable data of account
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AccountUpdate'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/Account'

            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Account"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
        delete:
            summary: Delete account
            operationId: deleteAccount
            tags:
                - accounts
            parameters:
                -   name: accountId
                    in: path
                    required: true
                    schema:
                        type: integer
            responses:
                '200':
                    description: ok
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
                default:
                    description: unexpected error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DefaultResponse"
...
components:
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
    schemas:
        Account:
            type: object
            required:
                - email
            properties:
                id:
                    type: integer
                    format: int32
                email:
                    type: string
                    format: email
        AccountCreate:
            type: object
            required:
                - email
                - password
            properties:
                email:
                    type: string
                    format: email
                password:
                    type: string
                    format: password
        AccountUpdate:
            type: object
            properties:
                email:
                    type: string
                    format: email
                password:
                    type: string
                    format: password
...
        DefaultResponse:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string